(PC-34815) feat: move and rename useSearchResults query
