+++ src/features/search/components/SearchResultsContent/SearchResultsContent.native.test.tsx
@@ -107,8 +107,8 @@ const initialSearchResults = {
 }
 
 const mockUseSearchResults = jest.fn(() => initialSearchResults)
-jest.mock('features/search/api/useSearchResults/useSearchResults', () => ({
-  useSearchResults: () => mockUseSearchResults(),
+jest.mock('features/search/queries/useSearchResultsQuery', () => ({
+  useSearchResultsQuery: () => mockUseSearchResults(),
 }))
 
 const mockedPlace: SuggestedPlace = {
