+++ src/features/search/components/SearchResultsContent/SearchResultsContent.tsx
@@ -11,7 +11,6 @@ import { useAuthContext } from 'features/auth/context/AuthContext'
 import { VenueMapLocationModal } from 'features/location/components/VenueMapLocationModal'
 import { getSearchNavConfig } from 'features/navigation/SearchStackNavigator/searchStackHelpers'
 import { PlaylistType } from 'features/offer/enums'
-import { useSearchResults } from 'features/search/api/useSearchResults/useSearchResults'
 import { AutoScrollSwitch } from 'features/search/components/AutoScrollSwitch/AutoScrollSwitch'
 import { FilterButton } from 'features/search/components/Buttons/FilterButton/FilterButton'
 import { SingleFilterButton } from 'features/search/components/Buttons/SingleFilterButton/SingleFilterButton'
@@ -33,6 +32,7 @@ import { DatesHoursModal } from 'features/search/pages/modals/DatesHoursModal/Da
 import { OfferDuoModal } from 'features/search/pages/modals/OfferDuoModal/OfferDuoModal'
 import { PriceModal } from 'features/search/pages/modals/PriceModal/PriceModal'
 import { VenueModal } from 'features/search/pages/modals/VenueModal/VenueModal'
+import { useSearchResultsQuery } from 'features/search/queries/useSearchResultsQuery'
 import { SearchView } from 'features/search/types'
 import { TabLayout } from 'features/venue/components/TabLayout/TabLayout'
 import { GeolocatedVenue } from 'features/venueMap/components/VenueMapView/types'
@@ -96,7 +96,7 @@ export const SearchResultsContent: React.FC = () => {
     venuesUserData,
     facets,
     offerVenues,
-  } = useSearchResults()
+  } = useSearchResultsQuery()
 
   const { disabilities } = useAccessibilityFiltersContext()
   const { searchState } = useSearch()
