+++ src/features/search/pages/SearchFilter/SearchFilter.web.test.tsx
@@ -22,8 +22,8 @@ jest.mock('libs/subcategories/useSubcategories', () => ({
 
 const mockHasNextPage = true
 const mockFetchNextPage = jest.fn()
-jest.mock('features/search/api/useSearchResults/useSearchResults', () => ({
-  useSearchResults: () => ({
+jest.mock('features/search/queries/useSearchResultsQuery', () => ({
+  useSearchResultsQuery: () => ({
     data: mockData,
     hits: [],
     nbHits: 0,
