+++ src/features/search/pages/ThematicSearch/ThematicSearch.native.test.tsx
@@ -85,8 +85,8 @@ const defaultUseSearchResults = {
   isFetchingNextPage: false,
 }
 const mockUseSearchResults = jest.fn(() => defaultUseSearchResults)
-jest.mock('features/search/api/useSearchResults/useSearchResults', () => ({
-  useSearchResults: () => mockUseSearchResults(),
+jest.mock('features/search/queries/useSearchResultsQuery', () => ({
+  useSearchResultsQuery: () => mockUseSearchResults(),
 }))
 
 const mockData = PLACEHOLDER_DATA
